Turned the SlideShow Components into Small Small Reusable Components
